Developed further on versions 1, 2, 3.
Hit project spec:
version 1.0 is made with nested if statements. Tell applicant why they
are non eligible.

version 2.0 is made with one if statement and logical operator and, and
does not tell applicant hwy they are not eligible

version 3.0 is made with no if statement and tells applicant why they
are eligible.